+++ Src/platform/ubuntu/pwm.cpp
@@ -5,13 +5,13 @@
 #include "periphery/pwm/pwm.hpp"
 #include <cstddef>
 
-void PwmPeriphery::set_duration(const PwmPin pwm_pin, uint32_t duration_us) {
+void PwmPeriphery::set_duration(const PwmPin pwm_pin, uint16_t duration_us) {
     (void)pwm_pin;
     (void)duration_us;
 
 }
 
-uint32_t PwmPeriphery::get_duration(PwmPin pwm_pin) {
+uint16_t PwmPeriphery::get_duration(PwmPin pwm_pin) {
     (void)pwm_pin;
     return 0;
 }
